# openapi.yaml
openapi: 3.1.0
info:
  title: SilverKey API
  version: 1.0.0
  description: Real estate AI platform API - comprehensive type definitions for frontend-backend contract
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://staging.silverkey.com
    description: Staging environment
  - url: https://api.silverkey.com
    description: Production
...
paths:
...
  /api/v1/report/list:
    get:
      tags:
        - Documents
      summary: List reports for current user
      operationId: listReports
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportsListResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: HTTP 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/{report_id}/download-url:
    get:
      tags:
        - Documents
      summary: Presigned download URL for report PDF
      operationId: getReportDownloadUrl
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: HTTP 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/{report_id}/view-url:
    get:
      tags:
        - Documents
      summary: Presigned view URL for report PDF
      operationId: getReportViewUrl
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewUrlResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: HTTP 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/{report_id}/view:
    get:
      tags:
        - Documents
      summary: Inline PDF bytes for report
      operationId: viewReportPdf
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTTP 200
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: HTTP 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/{report_id}:
    delete:
      tags:
        - Documents
      summary: Delete a report
      operationId: deleteReport
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteReportRequest'
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: HTTP 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/documents:
    get:
      tags:
        - Documents
      summary: Workflow documents for reports
      operationId: listReportWorkflowDocuments
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentsResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/document-library:
    get:
      tags:
        - Documents
      summary: Document library for current or scoped user
      operationId: getDocumentLibrary
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentLibraryResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/report/document-library/{library_item_id}:
    delete:
      tags:
        - Documents
      summary: Remove item from document library
      operationId: deleteDocumentLibraryItem
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: library_item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: HTTP 403
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: HTTP 404
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/upload/document:
    post:
      tags:
        - Documents
      summary: Secure document upload
      operationId: uploadSecureDocument
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                address:
                  type: string
      responses:
        '201':
          description: HTTP 201
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: HTTP 400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: HTTP 401
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: HTTP 500
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DeleteReportRequest:
      type: object
      properties:
        s3_key:
          type: string
          nullable: true
          description: Optional S3 key for the report to delete
        file_path:
          type: string
          nullable: true
          description: Optional legacy file path (alias of s3_key)
...
    DocumentLibraryResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/DocumentLibraryListItem'
              nullable: true
            count:
              type: integer
              nullable: true
              description: Number of items in this response page; `pagination.total` is the full library size
            pagination:
              $ref: '#/components/schemas/Pagination'
              nullable: true
...
    DocumentsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: '#/components/schemas/UploadedDocumentRecord'
              nullable: true
            count:
              type: integer
              nullable: true
            total:
              type: integer
              nullable: true
            limit:
              type: integer
              nullable: true
            offset:
              type: integer
              nullable: true
            hasMore:
              type: boolean
              nullable: true
...
    DownloadUrlResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            downloadUrl:
              type: string
              format: uri
              nullable: true
            expires_at:
              type: string
              nullable: true
...
    ErrorResponse:
      type: object
      description: |
        Standard error envelope for JSON APIs. `success` is always false; `error` is a stable
        code for clients, while `message` is safe to show to users. Use `error_id` when
        correlating with server logs; `field_errors` carries validation details when present.
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
          description: Machine-oriented error code or short error type (e.g. validation_error, GOOGLE_OAUTH_FAILED).
        message:
          description: Human-readable message when provided.
          anyOf:
            - type: string
            - type: 'null'
        error_id:
          description: Unique error tracking ID from SecureErrorHandler (short UUID prefix) for server-side log correlation.
          anyOf:
            - type: string
            - type: 'null'
        retryable:
          description: True when the client may retry (e.g. transient Google token refresh / network failures, 503-style cases).
          anyOf:
            - type: boolean
            - type: 'null'
        field_errors:
          description: |
            Per-field validation messages from SecureErrorHandler (sanitized string per field); some validators may emit multiple messages per field as an array.
          anyOf:
            - type: object
              additionalProperties:
                oneOf:
                  - type: string
                  - type: array
                    items:
                      type: string
            - type: 'null'
        validation_errors:
          description: |
            Optional alternate validation map when included via secure error additional_info. Values are usually strings or lists of strings (same shape as field_errors); additionalProperties stays open for legacy or nested payloads from some validators.
          anyOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        details:
          description: Additional context as plain text or structured map (e.g. legacy Marshmallow validate_request).
          anyOf:
            - type: string
            - type: object
              additionalProperties: true
            - type: 'null'
        status_code:
          description: Optional HTTP status echoed in the JSON body; the authoritative status is the response status line.
          anyOf:
            - type: integer
            - type: 'null'
        retry_after:
          description: Suggested delay in seconds before retrying (e.g. rate limit responses).
          anyOf:
            - type: integer
            - type: 'null'
        allowed_types:
          description: Allowed values hint (e.g. some file upload error responses).
          anyOf:
            - type: array
              items:
                type: string
            - type: 'null'
      examples:
        invalidCredentials:
          summary: Wrong email or password
          value:
            success: false
            error: INVALID_CREDENTIALS
            message: The provided email or password is incorrect
            error_id: a1b2c3d4
            retryable: false
            status_code: 401
            field_errors: null
            validation_errors: null
            details: null
            retry_after: null
            allowed_types: null
...
    ReportListItem:
      type: object
      description: |
        Entry returned by GET /api/v1/report/list — property PDF reports for the current user (camelCase fields as produced by the API). Not the same shape as `UploadedDocumentRecord`.
      required:
        - id
        - status
        - generatedAt
        - address
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - generating
            - completed
            - error
        generatedAt:
          type: integer
          description: Unix timestamp in seconds
        address:
          type: string
        pdfUrl:
          type: string
          format: uri
          nullable: true
        s3Key:
          type: string
          nullable: true
...
    ReportsListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            reports:
              type: array
              items:
                $ref: '#/components/schemas/ReportListItem'
              nullable: true
...
    SuccessResponse:
      type: object
      description: |
        Base success envelope extended via `allOf` across many routes. `success` is always
        true; `message` is optional human copy; `error` is normally null on success responses.
      required:
        - success
      properties:
        success:
          type: boolean
          enum:
            - true
          description: Always true for this schema; failed requests use ErrorResponse instead.
        message:
          type: string
          nullable: true
          description: Optional short status (e.g. OK or operation-specific confirmation).
        error:
          type: string
          nullable: true
          description: Reserved for symmetry with errors; keep null on success.
      example:
        success: true
        message: OK
        error: null
...
    UploadedDocumentRecord:
      type: object
      description: |
        Row from the `documents` table / GET /api/v1/report/documents — file storage and processing pipeline (upload, processing, report generation). Distinct from dashboard workflow docs and unified library rows.
      required:
        - document_record_kind
        - id
        - filename
        - file_path
        - status
        - user_id
      properties:
        document_record_kind:
          type: string
          const: pipeline
          description: Discriminator for OpenAPI tooling; always `pipeline` for documents-table / pipeline rows.
        id:
          type: string
        filename:
          type: string
        file_path:
          type: string
        status:
          type: string
          description: Pipeline state for the stored file (DB allows additional legacy values).
          enum:
            - uploaded
            - processing
            - processed
            - error
            - generating
        user_id:
          type: string
        created_at:
          type: string
          nullable: true
        updated_at:
          type: string
          nullable: true
        document_type:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
    UploadResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            file_id:
              type: string
              nullable: true
            file_url:
              type: string
              format: uri
              nullable: true
            filename:
              type: string
              nullable: true
            file_size:
              type: integer
              nullable: true
            content_type:
              type: string
              nullable: true
            document:
              $ref: '#/components/schemas/SecureUploadDocumentPayload'
              nullable: true
...
    ViewUrlResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            viewUrl:
              type: string
              format: uri
              nullable: true
            expires_at:
              type: string
              nullable: true
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly JWT access token (primary browser auth).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional Authorization header fallback (same JWT as session cookie).